pyproject.toml
[project]
name = "the-council"
version = "0.1.0"
description = "8 minds. 1 question. Infinite consequences. Multi-agent deliberation with K-ZERO Console."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Kyeong Sun Kim", email = "kyeong.sun.kim@snu.ac.kr"},
]
keywords = ["multi-agent", "llm", "deliberation", "simulation", "k-zero"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "plotly>=5.18.0",
    "dash>=2.14.0",
]

[project.urls]
Homepage = "https://github.com/kyeongsunkim/the-council"
Repository = "https://github.com/kyeongsunkim/the-council"

[project.scripts]
council = "runner.council_runner:main"
kzero = "runner.demiurge:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
